Deprecate Go plugin
