+++ go.mod
@@ -18,7 +18,6 @@ require (
 	github.com/lib/pq v1.3.0 // indirect
 	github.com/mattn/go-isatty v0.0.12 // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
-	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.7.0 // indirect
 	go.mozilla.org/sops/v3 v3.7.1
 	golang.org/x/text v0.3.4 // indirect
