--- test/exec-generate-multiple-resources.yaml
@@ -1,8 +1,0 @@
-apiVersion: viaduct.ai/v1
-kind: ksops-exec
-metadata:
-  name: ksops-exec-multiple-generator
-files:
-  - test/secret-A.enc.yaml
-  - test/secret-B.enc.yaml
-  - test/secret-C.enc.yaml
